ppt_agent/ppt_agent_adk/requirements.txt
streamlit>=1.28.0
google-adk>=0.1.0
python-pptx>=0.6.21
jinja2>=3.1.0
pyyaml>=6.0 